refactor(taxes): extract OwnerSection component
Move OwnerSection to src/pages/taxes/components/OwnerSection.tsx.
It imports OwnerBadge and ChecklistRow internally.
Taxes.tsx 859→767L.

Part of #181
